+++ index.html
@@ -1,5 +1,5 @@
 <!doctype html>
-<html>
+<html lang="en">
 	<head>
 		<meta charset=utf-8>
 		<meta name=description content="">
@@ -26,75 +26,71 @@
 				font-family: "Arial", Georgia, Serif;
 			}
 		</style>
+	</head>
+	<body>
+		<ax-comments id="demo-comments-container"></ax-comments>
 
 		<!-- Init jquery-comments -->
 		<script type="text/javascript">
-			$(function() {
-				var saveComment = function(data) {
-
-					// Convert pings to human readable format
-					$(Object.keys(data.pings)).each(function(index, userId) {
-					    var fullname = data.pings[userId];
-					    var pingText = '@' + fullname;
-					    data.content = data.content.replace(new RegExp('@' + userId, 'g'), pingText);
-					});
-
-					return data;
-				}
-				$('#comments-container').comments({
-					profilePictureURL: 'https://viima-app.s3.amazonaws.com/media/public/defaults/user-icon.png',
-					currentUserId: 1,
-					roundProfilePictures: true,
-					textareaRows: 1,
-					enableAttachments: true,
-					enableHashtags: true,
-					enablePinging: true,
-					scrollContainer: $(window),
-					searchUsers: function(term, success, error) {
-					    setTimeout(function() {
-					        success(usersArray.filter(function(user) {
-					            var containsSearchTerm = user.fullname.toLowerCase().indexOf(term.toLowerCase()) != -1;
-					            var isNotSelf = user.id != 1;
-					            return containsSearchTerm && isNotSelf;
-					        }));
-					    }, 500);
-					},
-					getComments: function(success, error) {
-						setTimeout(function() {
-							success(commentsArray);
-						}, 500);
-					},
-					postComment: function(data, success, error) {
-						setTimeout(function() {
-							success(saveComment(data));
-						}, 500);
-					},
-					putComment: function(data, success, error) {
-						setTimeout(function() {
-							success(saveComment(data));
-						}, 500);
-					},
-					deleteComment: function(data, success, error) {
-						setTimeout(function() {
-							success();
-						}, 500);
-					},
-					upvoteComment: function(data, success, error) {
-						setTimeout(function() {
-							success(data);
-						}, 500);
-					},
-					validateAttachments: function(attachments, callback) {
-						setTimeout(function() {
-							callback(attachments);
-						}, 500);
-					},
+			const saveComment = data => {
+				// Convert pings to human readable format
+				Object.keys(data.pings).forEach((userId) => {
+					const fullname = data.pings[userId];
+					const pingText = '@' + fullname;
+					data.content = data.content.replace(new RegExp('@' + userId, 'g'), pingText);
 				});
+
+				return data;
+			};
+
+			document.getElementById('demo-comments-container').options({
+				profilePictureURL: 'https://viima-app.s3.amazonaws.com/media/public/defaults/user-icon.png',
+				currentUserId: 1,
+				roundProfilePictures: true,
+				textareaRows: 1,
+				enableAttachments: true,
+				enableHashtags: true,
+				enablePinging: true,
+				searchUsers: (term, success, error) => {
+					setTimeout(() => {
+						success(usersArray.filter(user => {
+							const containsSearchTerm = user.fullname.toLowerCase().indexOf(term.toLowerCase()) !== -1;
+							const isNotSelf = user.id !== 1;
+							return containsSearchTerm && isNotSelf;
+						}));
+					}, 500);
+				},
+				getComments: (success, error) => {
+					setTimeout(() => {
+						success(commentsArray);
+					}, 500);
+				},
+				postComment: (data, success, error) => {
+					setTimeout(() => {
+						success(saveComment(data));
+					}, 500);
+				},
+				putComment: (data, success, error) => {
+					setTimeout(() => {
+						success(saveComment(data));
+					}, 500);
+				},
+				deleteComment: (data, success, error) => {
+					setTimeout(() => {
+						success();
+					}, 500);
+				},
+				upvoteComment: (data, success, error) => {
+					setTimeout(() => {
+						success(data);
+					}, 500);
+				},
+				validateAttachments: (attachments, callback) => {
+					setTimeout(() => {
+						callback(attachments);
+					}, 500);
+				},
 			});
 		</script>
-
-	</head>
-	<body>
-		<div id="comments-container"></div>
 	</body>
 </html>
