+++ src/comment-util.ts
@@ -1,0 +1,69 @@
+import {CommentsOptions} from './comments-options';
+import {CommentsById} from './comments-by-id';
+import {SpinnerFactory} from './subcomponent/spinner-factory';
+import {CommentsProvider, OptionsProvider, ServiceProvider} from './provider';
+import {isNil} from './util';
+
+export class CommentUtil {
+
+    private readonly options: CommentsOptions;
+    private readonly commentsById: CommentsById;
+    private readonly spinnerFactory: SpinnerFactory;
+
+    constructor(private readonly container: HTMLDivElement) {
+        this.options = OptionsProvider.get(container)!;
+        this.commentsById = CommentsProvider.get(container)!;
+        this.spinnerFactory = ServiceProvider.get(container, SpinnerFactory);
+    }
+
+    getComments(): Record<string, any>[] {
+        return Object.keys(this.commentsById).map(id => this.commentsById[id]);
+    }
+
+    getChildComments(parentId: string): Record<string, any>[] {
+        return this.getComments().filter(comment => comment.parent === parentId);
+    }
+
+    getAttachments(): Record<string, any>[] {
+        return this.getComments().filter(comment => comment.hasAttachments());
+    }
+
+    getOutermostParent(directParentId: string) {
+        let parentId = directParentId;
+        let parentComment;
+        do {
+            parentComment = this.commentsById[parentId];
+            parentId = parentComment.parent;
+        } while (!isNil(parentComment.parent));
+        return parentComment;
+    }
+
+    removeComment(commentId: string, onCommentRemoved: (parentEl: HTMLElement) => void): void {
+        const commentModel = this.commentsById[commentId];
+
+        // Remove child comments recursively
+        const childComments: Record<string, any>[] = this.getChildComments(commentModel.id);
+        childComments.forEach(childComment => {
+            this.removeComment(childComment.id, onCommentRemoved);
+        });
+
+        // Update the child array of outermost parent
+        if (commentModel.parent) {
+            const outermostParent = this.getOutermostParent(commentModel.parent);
+            const indexToRemove = outermostParent.childs.indexOf(commentModel.id);
+            outermostParent.childs.splice(indexToRemove, 1);
+        }
+
+        // Remove the comment from data model
+        delete this.commentsById[commentId];
+
+        const commentElement: HTMLElement = this.container.querySelector(`li.comment[data-id="${commentId}"]`)!;
+        const parentEl = commentElement.parents('li.comment').last();
+
+        // Remove the element
+        commentElement.remove();
+
+        // Execute callback
+        onCommentRemoved(parentEl);
+    }
+}
